chore: add benchmark mode
Benchmark mode avoids actually sending out DNS queries upstream but
answers with a pre-defined, empty DNS response.

We do this in order to eliminate any volatility in upstream DNS and
network problems to make our benchmarks more meaningful, before we focus
on implementing performance improvements.

Of course, real-world benchmarks should not just exclude half of the
network stack work, but we need a more stable and consistent
benchmarking environment for that, ie. I don't trust GitHub action
runners with this.