fix(resolution): use separate ephemeral port to relay responses
Generally, sharing a common socket to send the response from
upstream to the user would seem sensible, so to avoid binding
new sockets upon serving requests dynamically.

However, with this sharing approach, concurrent DNS requests
seem to cause requests being served out of order and individual
requests being answered with the response from another request
and vice-versa.

This can be observed when running additional dig queries in
parallel to a benchmark invocation and dig reporting a mismatch
in request ID at the DNS protocol level.

Rerunning this experiment with this change seems to fix that.